refactor: コンポーネント階層を今のやり方に変更(#10)
