+++ src/components/Layout.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
 import Head from 'next/head';
 import { css } from '@emotion/react';
-import Header from '@/components/Header';
-import Footer from '@/components/Footer';
+import Header from '@/components/common/Header';
+import Footer from '@/components/common/Footer';
 
 type Props = {
   pageName?: string;
