+++ src/components/common/Alert/index.ts
@@ -1,0 +1,3 @@
+import Alert from './Alert';
+
+export default Alert;
