+++ src/components/common/Header/index.ts
@@ -1,0 +1,3 @@
+import Header from './Header';
+
+export default Header;
