+++ src/components/common/RandomButton/index.ts
@@ -1,0 +1,3 @@
+import RandomButton from './RandomButton';
+
+export default RandomButton;
