+++ src/components/common/Skeleton/index.ts
@@ -1,0 +1,3 @@
+import Skeleton from './Skeleton';
+
+export default Skeleton;
