+++ src/components/model/Quote/QuoteAuthorLinkCard/index.ts
@@ -1,0 +1,3 @@
+import QuoteAuthorLinkCard from './QuoteAuthorLinkCard';
+
+export default QuoteAuthorLinkCard;
