+++ src/components/model/Quote/QuoteBlock/QuoteBlock.tsx
