+++ src/components/model/Quote/QuoteBlock/index.ts
@@ -1,0 +1,3 @@
+import QuoteBlock from './QuoteBlock';
+
+export default QuoteBlock;
