+++ src/components/model/Quote/QuoteBlockList/index.ts
@@ -1,0 +1,3 @@
+import QuoteBlockList from './QuoteBlockList';
+
+export default QuoteBlockList;
