+++ src/components/page/AuthorQuotes/index.ts
@@ -1,0 +1,3 @@
+import AuthorQuotes from './AuthorQuotes.page';
+
+export default AuthorQuotes;
