+++ src/components/page/Home/index.ts
@@ -1,0 +1,3 @@
+import Home from './Home.page';
+
+export default Home;
